+++ app/src/main/java/com/saradabar/cpadcustomizetool/util/DchaUtilServiceUtil.java
@@ -20,6 +20,11 @@ public class DchaUtilServiceUtil {
 
     public void clearDefaultPreferredApp(String packageName, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 iDchaUtilService.clearDefaultPreferredApp(packageName);
                 listener.onResult(true);
@@ -31,6 +36,11 @@ public class DchaUtilServiceUtil {
 
     public void copyFile(String srcFilePath, String dstFilePath, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.copyFile(srcFilePath, dstFilePath));
             } catch (RemoteException e) {
@@ -41,9 +51,13 @@ public class DchaUtilServiceUtil {
 
     public void copyDirectory(String srcFilePath, String dstFilePath, boolean makeTopDir, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
-                iDchaUtilService.copyDirectory(srcFilePath, dstFilePath, makeTopDir);
-                listener.onResult(true);
+                listener.onResult(iDchaUtilService.copyDirectory(srcFilePath, dstFilePath, makeTopDir));
             } catch (RemoteException e) {
                 listener.onResult(e.getMessage());
             }
@@ -52,6 +66,11 @@ public class DchaUtilServiceUtil {
 
     public void deleteFile(String path, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.deleteFile(path));
             } catch (RemoteException e) {
@@ -62,6 +81,11 @@ public class DchaUtilServiceUtil {
 
     public void existsFile(String path, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.existsFile(path));
             } catch (RemoteException e) {
@@ -72,6 +96,11 @@ public class DchaUtilServiceUtil {
 
     public void getCanonicalExternalPath(String linkPath, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.getCanonicalExternalPath(linkPath));
             } catch (RemoteException e) {
@@ -82,6 +111,11 @@ public class DchaUtilServiceUtil {
 
     public void getDisplaySize(Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.getDisplaySize());
             } catch (RemoteException e) {
@@ -92,6 +126,11 @@ public class DchaUtilServiceUtil {
 
     public void getLcdSize(Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.getLcdSize());
             } catch (RemoteException e) {
@@ -102,6 +141,11 @@ public class DchaUtilServiceUtil {
 
     public void getUserCount(Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.getUserCount());
             } catch (RemoteException e) {
@@ -112,6 +156,11 @@ public class DchaUtilServiceUtil {
 
     public void hideNavigationBar(boolean hide, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 iDchaUtilService.hideNavigationBar(hide);
                 listener.onResult(true);
@@ -123,6 +172,11 @@ public class DchaUtilServiceUtil {
 
     public void listFiles(String path, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.listFiles(path));
             } catch (RemoteException e) {
@@ -133,6 +187,11 @@ public class DchaUtilServiceUtil {
 
     public void makeDir(String path, String dirname, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.makeDir(path, dirname));
             } catch (RemoteException e) {
@@ -143,6 +202,11 @@ public class DchaUtilServiceUtil {
 
     public void sdUnmount(Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 iDchaUtilService.sdUnmount();
                 listener.onResult(true);
@@ -154,6 +218,11 @@ public class DchaUtilServiceUtil {
 
     public void setForcedDisplaySize(int width, int height, Listener listener) {
         new IDchaUtilTask().execute(mContext, iDchaUtilService -> {
+            if (iDchaUtilService == null) {
+                listener.onResult(false);
+                return;
+            }
+
             try {
                 listener.onResult(iDchaUtilService.setForcedDisplaySize(width, height));
             } catch (RemoteException e) {
